--- jee-in/이론 실습/src/ch02_BasicDataStructure/ch02_1_Array/Quiz/q08_dateClass.java
@@ -1,24 +1,0 @@
-package ch02_BasicDataStructure.ch02_1_Array.Quiz;
-
-/***************************************************
- * Q8. 년월일을 필드(멤버 변수)로 갖는 클래스 만들기
- ****************************************************/
-
-public class q08_dateClass {
-
-  int y; // 년
-  int m; // 월
-  int d; // 일
-
-  public q08_dateClass (int y, int m, int d) {
-    return 
-  }
-
-  public q08_dateClass after (int n) {
-    return 
-  }
-
-  public q08_dateClass before (int n) {
-    return 
-  }
-}
