+++ jee-in/이론 실습/src/ch02_basic/part1_array/c01_intArray.java
@@ -1,4 +1,4 @@
-package ch02_BasicDataStructure.ch02_1_Array;
+package ch02_basic.part1_array;
 
 /***************************************************
  * 숫자 배열 기초
