+++ jee-in/이론 실습/src/ch02_basic/part1_array/c08_PrimeNumber03.java
@@ -1,4 +1,4 @@
-package ch02_BasicDataStructure.ch02_1_Array;
+package ch02_basic.part1_array;
 
 /********************************************
  * 1000 이하의 소수 나열하기
