+++ jee-in/이론 실습/src/ch02_basic/part1_array/c09_CloneArray.java
@@ -1,4 +1,4 @@
-package ch02_BasicDataStructure.ch02_1_Array;
+package ch02_basic.part1_array;
 
 /********************************************
  * 배열 복사하기
